pyproject.toml
[project]
name = "code-translator-green-agent"
version = "0.1.0"
description = "Green Agent for Code Translator"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "uvicorn",
    "httpx",
    "google-genai",
    "pydantic",
    "google-adk[a2a]"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]